+++ CHANGELOG.md
@@ -1,5 +1,7 @@
+## 0.3.1
+    added html output format
 ## 0.2.1
-    now you don't need @diInject annotation. Analyzer will get info(params) from constructors
+    now you don't need `@diInject` annotation. Analyzer will get info(params) from constructors
 ## 0.1.1
     revert
 ## 0.0.4